Upgrade react-hot-loader
