+++ package.json
@@ -37,6 +37,7 @@
     "@types/react-dnd": "^2.0.34",
     "@types/react-dnd-html5-backend": "^2.1.8",
     "@types/react-dom": "^15.5.4",
+    "@types/react-hot-loader": "^3.0.4",
     "@types/uuid": "^3.4.2",
     "chai": "^4.1.2",
     "css-loader": "^0.28.7",
@@ -53,7 +54,7 @@
     "react": "^16.0.0",
     "react-addons-css-transition-group": "^15.6.2",
     "react-dom": "^16.0.0",
-    "react-hot-loader": "^1.3.1",
+    "react-hot-loader": "3.0.0-beta.7",
     "replace": "v0.3.0",
     "source-map-loader": "^0.2.1",
     "style-loader": "^0.18.2",
@@ -65,7 +66,6 @@
     "webpack-dev-server": "^1.16.2"
   },
   "dependencies": {
-    "@types/react-hot-loader": "^3.0.4",
     "classnames": "^2.2.5",
     "immutability-helper": "^2.4.0",
     "lodash": "^4.17.2",
