+++ src/index.ts
@@ -63,21 +63,15 @@ export {
   getPathToCorner,
   isParent,
 } from './mosaicUtilities';
-export { MosaicWindow, MosaicWindowFactory, MosaicWindowProps } from './window/MosaicWindow';
+export { MosaicWindow, MosaicWindowFactory, MosaicWindowProps } from './MosaicWindow';
 export {
   createDefaultToolbarButton,
-  DEFAULT_CONTROLS_WITH_CREATION,
-  DEFAULT_CONTROLS_WITHOUT_CREATION,
-  ExpandButton,
-  ExpandButtonFactory,
   MosaicButtonProps,
-  RemoveButton,
-  RemoveButtonFactory,
-  ReplaceButton,
-  ReplaceButtonFactory,
-  Separator,
-  SeparatorFactory,
-  SplitButton,
-  SplitButtonFactory,
-} from './window/defaultToolbarControls';
+} from './buttons/MosaicButton';
 export { MosaicZeroState, MosaicZeroStateFactory, MosaicZeroStateProps } from './MosaicZeroState';
+export { Separator, SeparatorFactory } from './buttons/Separator';
+export { ExpandButton, ExpandButtonFactory } from './buttons/ExpandButton';
+export { ReplaceButton, ReplaceButtonFactory } from './buttons/ReplaceButton';
+export { SplitButton, SplitButtonFactory } from './buttons/SplitButton';
+export { RemoveButton, RemoveButtonFactory } from './buttons/RemoveButton';
+export { DEFAULT_CONTROLS_WITH_CREATION, DEFAULT_CONTROLS_WITHOUT_CREATION } from './buttons/defaultToolbarControls';
